+++ generators/java/sdk/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.2] - 2024-02-21
+- Fix: File upload endpoints no longer fail to compile because the reference to 
+  the mime type variable is present.  
+  
+  ```java
+  // Code that failed to compile
+  String fileMimeType = Files.probeContentType(file.toPath());
+  MediaType fileMediaType = fileMimeType != null ? MediaType.parse(mimeType) : null; // mimeType undefined
+  // Code that now compiles
+  MediaType fileMediaType = fileMimeType != null ? MediaType.parse(fileMimeType) : null;
+  ```
+
 ## [0.8.1] - 2024-02-14
 - Feature: The RequestOptions object now supports configuring an optional timeout to apply per-request.
   ```java
@@ -13,7 +25,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   //  RequestOptions ro = RequestOptions.builder().timeout(2, TimeUnit.MINUTES).build();
   client.films.list(ro);
   ```
-
 
 ## [0.8.0] - 2024-02-11
 - Feature: The SDK generator now supports whitelabelling. When this is turned on,